Lift state to CreateTrackContainer. Add countdown timer to UI
