+++ src/components/tracks/BlockRow.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 import Block from './Block';
-import { Colors } from '../../styles/colors';
-
 
 const rowStyle = {
   display: 'flex',
@@ -9,15 +7,14 @@ const rowStyle = {
   height: '20%',
   borderRadius: '5px',
   overflow: 'hidden',
-  // backgroundColor: 'red'
 }
 
 const BlockRow = (props) => {
-  const { color, track } = props;
+  const { color, inProgress, state, onCountdownFinished } = props;
   const loops = [1];
   return (
     <div style={rowStyle}>
-      {loops.map( (loop, idx) => <Block color={color} loop={loop} key={idx} progress={0} />)}
+      {loops.map((loop, idx) => <Block state={state} onCountdownFinished={onCountdownFinished} inProgress={inProgress} color={color} loop={loop} loopIndex={idx} key={idx} progress={0} />)}
     </div>
   );
 }
